chore: huge reorg + slimdown of express config + renaming
