+++ packages/backend/package.json
@@ -10,9 +10,7 @@
   "license": "MIT",
   "scripts": {
     "build": "webpack --watch --mode development --config webpack.config.js",
-    "build:prod": "webpack --mode production --config webpack.config.js",
-    "start": "pm2-dev --raw dist/server.js",
-    "start:prod": "pm2-runtime dist/server.js",
+    "start": "nodemon --ignore ./dist --exec babel-node ./src/index.js",
     "test": "mocha test",
     "test:services": "mocha test/services",
     "test:providers": "mocha test/providers",
@@ -42,6 +40,7 @@
     "big-integer": "^1.6.48",
     "bull": "^3.14.0",
     "cookie-parser": "^1.4.4",
+    "core-js": "^3.9.0",
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
     "express": "^4.16.1",
@@ -52,10 +51,12 @@
     "mongodb": "^3.5.4",
     "morgan": "^1.9.1",
     "node-cron": "^2.0.3",
+    "nodemon": "^2.0.7",
     "numeral": "^2.0.6",
     "nyc": "^15.0.0",
     "pm2": "^4.3.1",
     "query-string": "^6.11.1",
+    "regenerator-runtime": "^0.13.7",
     "stripe": "^8.33.0",
     "strong-cryptor": "^2.2.0",
     "uuid": "^8.3.0"
@@ -74,7 +75,6 @@
     "chai": "^4.2.0",
     "chai-http": "^4.3.0",
     "debug": "^2.6.9",
-    "dotenv-webpack": "^1.7.0",
     "eslint": "^6.8.0",
     "eslint-config-prettier": "^7.2.0",
     "eslint-loader": "^3.0.3",
@@ -83,10 +83,6 @@
     "mocha": "^8.1.3",
     "prettier": "^2.2.1",
     "sinon": "^9.0.1",
-    "subpackage": "^1.1.0",
-    "webpack": "^4.41.6",
-    "webpack-cli": "^3.3.11",
-    "webpack-dev-middleware": "^3.7.2",
-    "webpack-node-externals": "^1.7.2"
+    "subpackage": "^1.1.0"
   }
 }
