+++ packages/backend/src/_routes.js
@@ -2,11 +2,9 @@ import Index from "./apis/IndexApi";
 import User from "./apis/UserApi";
 import Account from "./apis/AccountApi";
 import SecurityQuestion from "./apis/SecurityQuestionApi";
-import Payment from "./apis/PaymentApi";
 import Application from "./apis/ApplicationApi";
 import Node from "./apis/NodeApi";
 import Network from "./apis/NetworkApi";
-import Checkout from "./apis/CheckoutApi";
 
 /**
  * @param {object} expressApp Express application object.
@@ -24,9 +22,6 @@ export function configureRoutes(expressApp) {
   // Security Questions API
   expressApp.use("/api/security_questions", SecurityQuestion);
 
-  // Payments API
-  expressApp.use("/api/payments", Payment);
-
   // Applications API
   expressApp.use("/api/applications", Application);
 
@@ -35,7 +30,4 @@ export function configureRoutes(expressApp) {
 
   // Network API
   expressApp.use("/api/network", Network);
-
-  // Checkout API
-  expressApp.use("/api/checkout", Checkout);
 }
