+++ packages/backend/src/excluded-paths.js
@@ -1,0 +1,17 @@
+export const excludedPaths = [
+  "/api/users/validate-token",
+  "/api/users/exists",
+  "/api/users/auth/login",
+  "/api/users/auth/signup",
+  "/api/users/auth/providers",
+  "/api/users/auth/providers/login",
+  "/api/users/auth/resend-signup-email",
+  "/api/users/auth/is-validated",
+  "/api/users/auth/reset-password",
+  "/api/users/auth/send-reset-password-email",
+  "/api/users/auth/verify-captcha",
+  "/api/security_questions/user/validate-answers",
+  "/api/security_questions/user/all",
+  "/api/security_questions/answered",
+  "login",
+];
