+++ packages/backend/src/providers/data/db/Index.js
@@ -1,9 +1,7 @@
 import MongoDBAdapter from "./MongoDBAdapter";
-import { Configurations } from "../../../_configuration";
+import env from "environment";
 
-const DEFAULT_DB_PROVIDER = new MongoDBAdapter(
-  Configurations.persistence.default
-);
+const DEFAULT_DB_PROVIDER = new MongoDBAdapter(env("persistence").default);
 
 /**
  * @returns {MongoDBAdapter} The default mongo adapter provider.
