--- packages/backend/src/providers/payment/README.md
@@ -1,1 +1,0 @@
-# Payment Providers
