+++ packages/backend/src/services/PocketService.js
@@ -16,10 +16,12 @@ import {
   UnlockedAccount,
   PocketAAT,
 } from "@pokt-network/pocket-js";
-import { Configurations } from "../_configuration";
+import env from "environment";
 import { PocketNetworkError } from "../models/Exceptions";
 
-const POCKET_NETWORK_CONFIGURATION = Configurations.pocket_network;
+const POCKET_NETWORK_CONFIGURATION = env("pocket_network");
+
+console.log(POCKET_NETWORK_CONFIGURATION.dispatchers);
 
 const POCKET_CONFIGURATION = new Configuration(
   POCKET_NETWORK_CONFIGURATION.max_dispatchers,
