+++ packages/backend/test/services/AccountService.spec.js
@@ -1,10 +1,11 @@
-import {before, describe, it} from "mocha";
-import {assert} from "chai";
-import {configureTestService} from "../setupTests";
+import { before, describe, it } from "mocha";
+import { assert } from "chai";
+import { configureTestService } from "../setupTests";
 import AccountService from "../../src/services/AccountService";
 
 /** @type {string} */
-const APPLICATION_ACCOUNT_IN_NETWORK = process.env.TEST_APPLICATION_ACCOUNT_IN_NETWORK;
+const APPLICATION_ACCOUNT_IN_NETWORK =
+  process.env.TEST_APPLICATION_ACCOUNT_IN_NETWORK;
 
 const accountService = new AccountService();
 
@@ -13,12 +14,12 @@ before(() => {
 });
 
 describe("AccountService", () => {
-
   if (APPLICATION_ACCOUNT_IN_NETWORK) {
     describe("getBalance", () => {
       it("Expect an account balance", async () => {
-
-        const balance = await accountService.getBalance(APPLICATION_ACCOUNT_IN_NETWORK);
+        const balance = await accountService.getBalance(
+          APPLICATION_ACCOUNT_IN_NETWORK
+        );
 
         assert.isDefined(balance);
         assert.isNumber(balance);
