+++ packages/backend/test/services/PaymentService.spec.js
@@ -1,10 +1,13 @@
-import {before, describe, it} from "mocha";
+import { before, describe, it } from "mocha";
 import "chai/register-should";
-import {Configurations} from "../../src/_configuration";
-import {PaymentCurrencies, PaymentTypes} from "../../src/providers/payment/BasePaymentProvider";
+import { Configurations } from "../../src/_configuration";
+import {
+  PaymentCurrencies,
+  PaymentTypes,
+} from "../../src/providers/payment/BasePaymentProvider";
 import PaymentService from "../../src/services/PaymentService";
-import {configureTestService} from "../setupTests";
-import {PaymentHistory, PaymentMethod} from "../../src/models/Payment";
+import { configureTestService } from "../setupTests";
+import { PaymentHistory, PaymentMethod } from "../../src/models/Payment";
 
 const paymentService = new PaymentService();
 
@@ -13,10 +16,8 @@ before(() => {
 });
 
 describe("PaymentService", () => {
-
   describe("getAvailableCurrencies", () => {
     it("List of available currencies", () => {
-
       const currencies = paymentService.getAvailableCurrencies();
 
       // eslint-disable-next-line no-undef
@@ -28,7 +29,6 @@ describe("PaymentService", () => {
 
   describe("savePaymentMethod", () => {
     it("Save a payment method without error", async () => {
-
       const paymentMethodData = {
         user: "tester@testing.com",
         id: "paymentMethodTestID",
@@ -37,9 +37,9 @@ describe("PaymentService", () => {
           address: {
             line1: "Test billing detail",
             postal_code: "Test postal code",
-            country: "TE"
-          }
-        }
+            country: "TE",
+          },
+        },
       };
 
       const saved = await paymentService.savePaymentMethod(paymentMethodData);
@@ -53,7 +53,6 @@ describe("PaymentService", () => {
 
   describe("paymentMethodExists", () => {
     it("Expect a true value", async () => {
-
       const paymentMethodData = {
         user: "tester@testing.com",
         id: "paymentMethodTestID",
@@ -62,12 +61,14 @@ describe("PaymentService", () => {
           address: {
             line1: "Test billing detail",
             postal_code: "Test postal code",
-            country: "TE"
-          }
-        }
+            country: "TE",
+          },
+        },
       };
 
-      const paymentMethod = PaymentMethod.createPaymentMethod(paymentMethodData);
+      const paymentMethod = PaymentMethod.createPaymentMethod(
+        paymentMethodData
+      );
       const exists = await paymentService.paymentMethodExists(paymentMethod);
 
       exists.should.be.true;
@@ -76,7 +77,6 @@ describe("PaymentService", () => {
 
   describe("deletePaymentMethod", () => {
     it("Expect a true value", async () => {
-
       const id = "paymentMethodTestID";
       const deleted = await paymentService.deletePaymentMethod(id);
 
@@ -86,19 +86,25 @@ describe("PaymentService", () => {
 
   describe("savePaymentHistory", () => {
     it("Save a payment history without error", async () => {
-
       const user = "tester@testing.com";
       const createdDate = "2020-05-08T00:00:00.000Z";
       const paymentID = "paymentMethodTestID";
       const currency = "usd";
       const amount = 1000;
       const item = {
-        "account": "b9628a13220f049b93fdefe4fb9ca2bca10fe460",
-        "name": "My App",
-        "pokt": "346"
+        account: "b9628a13220f049b93fdefe4fb9ca2bca10fe460",
+        name: "My App",
+        pokt: "346",
       };
 
-      const saved = await paymentService.savePaymentHistory(createdDate, paymentID, currency, amount, item, user);
+      const saved = await paymentService.savePaymentHistory(
+        createdDate,
+        paymentID,
+        currency,
+        amount,
+        item,
+        user
+      );
 
       // eslint-disable-next-line no-undef
       should.exist(saved);
@@ -109,7 +115,6 @@ describe("PaymentService", () => {
 
   describe("paymentHistoryExists", () => {
     it("Expect a true value", async () => {
-
       const paymentHistoryData = {
         user: "tester@testing.com",
         createdDate: "2020-05-08T00:00:00.000Z",
@@ -120,11 +125,13 @@ describe("PaymentService", () => {
           account: "b9628a13220f049b93fdefe4fb9ca2bca10fe460",
           name: "My App",
           maxRelay: "3243",
-          pokt: "346"
-        }
+          pokt: "346",
+        },
       };
 
-      const paymentHistory = PaymentHistory.createPaymentHistory(paymentHistoryData);
+      const paymentHistory = PaymentHistory.createPaymentHistory(
+        paymentHistoryData
+      );
       const exists = await paymentService.paymentHistoryExists(paymentHistory);
 
       exists.should.be.true;
@@ -133,11 +140,13 @@ describe("PaymentService", () => {
 
   describe("getPaymentHistory", () => {
     it("Expect a list of payments", async () => {
-
       const user = "tester@testing.com";
       const limit = 10;
 
-      const paymentHistory = await paymentService.getPaymentHistory(user, limit);
+      const paymentHistory = await paymentService.getPaymentHistory(
+        user,
+        limit
+      );
 
       paymentHistory.should.be.an("array");
       paymentHistory.length.should.be.greaterThan(0);
@@ -146,14 +155,19 @@ describe("PaymentService", () => {
 
   describe("getPaymentHistory with fromDate toDate", () => {
     it("Expect a record of a payment", async () => {
-
       const user = "tester@testing.com";
       const limit = 10;
       const offset = 0;
       const fromDate = "2020-05-08";
       const toDate = "2020-05-08";
 
-      const paymentHistory = await paymentService.getPaymentHistory(user, limit, offset, fromDate, toDate);
+      const paymentHistory = await paymentService.getPaymentHistory(
+        user,
+        limit,
+        offset,
+        fromDate,
+        toDate
+      );
 
       paymentHistory.should.be.an("array");
       paymentHistory.length.should.be.equal(1);
@@ -162,10 +176,11 @@ describe("PaymentService", () => {
 
   describe("getPaymentHistoryByID", () => {
     it("Expect a payment", async () => {
-
       const paymentMethodID = "paymentMethodTestID";
 
-      const paymentHistory = await paymentService.getPaymentFromHistory(paymentMethodID);
+      const paymentHistory = await paymentService.getPaymentFromHistory(
+        paymentMethodID
+      );
 
       paymentHistory.should.be.an("object");
     });
@@ -173,7 +188,6 @@ describe("PaymentService", () => {
 
   describe("markPaymentAsSuccess", () => {
     it("Expect a true value", async () => {
-
       const paymentHistoryData = {
         user: "tester@testing.com",
         paymentID: "paymentMethodTestID",
@@ -183,18 +197,23 @@ describe("PaymentService", () => {
           address: {
             line1: "Test billing detail",
             postal_code: "Test postal code",
-            country: "TE"
-          }
-        }
+            country: "TE",
+          },
+        },
       };
 
-      const exists = await paymentService.markPaymentAsSuccess(paymentHistoryData);
+      const exists = await paymentService.markPaymentAsSuccess(
+        paymentHistoryData
+      );
 
       exists.should.be.equal(true);
     });
   });
 
-  if (Configurations.payment.test.client_id && Configurations.payment.test.client_secret) {
+  if (
+    Configurations.payment.test.client_id &&
+    Configurations.payment.test.client_secret
+  ) {
     describe("createPaymentIntent for apps that don’t require authentication", () => {
       it.skip("Create a Payment with amount, currency and type successfully", async () => {
         const type = PaymentTypes.card;
@@ -202,11 +221,16 @@ describe("PaymentService", () => {
         const item = {
           account: "b9628a13220f049b93fdefe4fb9ca2bca10fe460",
           name: "My App",
-          pokt: "346"
+          pokt: "346",
         };
         const amount = 90;
 
-        const paymentResult = await paymentService.createPocketPaymentIntentForApps(type, currency, item, amount);
+        const paymentResult = await paymentService.createPocketPaymentIntentForApps(
+          type,
+          currency,
+          item,
+          amount
+        );
 
         // eslint-disable-next-line no-undef
         should.exist(paymentResult);
@@ -223,11 +247,16 @@ describe("PaymentService", () => {
         const item = {
           account: "b9628a13220f049b93fdefe4fb9ca2bca10fe460",
           name: "My Node",
-          pokt: "346"
+          pokt: "346",
         };
         const amount = 90;
 
-        const paymentResult = await paymentService.createPocketPaymentIntentForNodes(type, currency, item, amount);
+        const paymentResult = await paymentService.createPocketPaymentIntentForNodes(
+          type,
+          currency,
+          item,
+          amount
+        );
 
         // eslint-disable-next-line no-undef
         should.exist(paymentResult);
