--- packages/backend/webpack.config.js
@@ -1,49 +1,0 @@
-const path = require("path");
-const nodeExternals = require("webpack-node-externals");
-const Dotenv = require("dotenv-webpack");
-const webpack = require("webpack");
-
-module.exports = {
-  mode: "development",
-  entry: {
-    server: "./src/App.js",
-  },
-  output: {
-    globalObject: "this",
-    path: path.join(__dirname, "dist"),
-    publicPath: "/",
-    filename: "[name].js",
-  },
-  target: "web",
-  devtool: "inline-source-map",
-  node: {
-    // Need this when working with express, otherwise the build fails
-    __dirname: false, // if you don't put this is, __dirname
-    __filename: false, // and __filename return blank or /
-  },
-  externals: [nodeExternals()], // Need this to avoid error when working with Express
-  module: {
-    rules: [
-      {
-        enforce: "pre",
-        test: /\.js$/,
-        exclude: /node_modules/,
-        loader: "eslint-loader",
-        options: {
-          emitWarning: true,
-          failOnError: false,
-          failOnWarning: false,
-        },
-      },
-      {
-        // Transpiles ES6-8 into ES5
-        test: /\.js$/,
-        exclude: /node_modules/,
-        use: {
-          loader: "babel-loader",
-        },
-      },
-    ],
-  },
-  plugins: [new webpack.NoEmitOnErrorsPlugin(), new Dotenv()],
-};
